Move Engines classes to PhpMyAdmin namespace
