+++ libraries/classes/Engines/Bdb.php
@@ -5,7 +5,7 @@
  *
  * @package PhpMyAdmin-Engines
  */
-namespace PMA\libraries\engines;
+namespace PhpMyAdmin\Engines;
 
 use PhpMyAdmin\StorageEngine;
