+++ libraries/classes/Engines/Binlog.php
@@ -1,20 +1,20 @@
 <?php
 /* vim: set expandtab sw=4 ts=4 sts=4: */
 /**
- * The performance schema storage engine
+ * The binary log storage engine
  *
  * @package PhpMyAdmin-Engines
  */
-namespace PMA\libraries\engines;
+namespace PhpMyAdmin\Engines;
 
 use PhpMyAdmin\StorageEngine;
 
 /**
- * The performance schema storage engine
+ * The binary log storage engine
  *
  * @package PhpMyAdmin-Engines
  */
-class PerformanceSchema extends StorageEngine
+class Binlog extends StorageEngine
 {
     /**
      * Returns string with filename for the MySQL helppage
@@ -24,6 +24,7 @@ class PerformanceSchema extends StorageEngine
      */
     public function getMysqlHelpPage()
     {
-        return 'performance-schema';
+        return 'binary-log';
     }
 }
+
