+++ libraries/classes/Engines/Innobase.php
@@ -1,18 +1,17 @@
 <?php
 /* vim: set expandtab sw=4 ts=4 sts=4: */
 /**
- * The BerkeleyDB storage engine
+ * The Innobase storage engine
  *
  * @package PhpMyAdmin-Engines
  */
-namespace PMA\libraries\engines;
+namespace PhpMyAdmin\Engines;
 
 /**
- * This is same as BDB
+ * The Innobase storage engine
  *
  * @package PhpMyAdmin-Engines
  */
-class Berkeleydb extends Bdb
+class Innobase extends Innodb
 {
 }
-
