+++ test/classes/engines/InnodbTest.php
@@ -9,13 +9,13 @@
  * Include to test.
  */
 
-use PMA\libraries\engines\Innodb;
+use PhpMyAdmin\Engines\Innodb;
 
 require_once 'libraries/database_interface.inc.php';
 require_once 'test/PMATestCase.php';
 
 /**
- * Tests for PMA\libraries\engines\Innodb
+ * Tests for PhpMyAdmin\Engines\Innodb
  *
  * @package PhpMyAdmin-test
  */
