+++ test/classes/engines/MrgMyisamTest.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Tests for PMA\libraries\engines\MrgMyisam
+ * Tests for PhpMyAdmin\Engines\MrgMyisam
  *
  * @package PhpMyAdmin-test
  */
@@ -9,13 +9,13 @@
  * Include to test.
  */
 
-use PMA\libraries\engines\MrgMyisam;
+use PhpMyAdmin\Engines\MrgMyisam;
 
 require_once 'libraries/database_interface.inc.php';
 require_once 'test/PMATestCase.php';
 
 /**
- * Tests for PMA\libraries\engines\MrgMyisam
+ * Tests for PhpMyAdmin\Engines\MrgMyisam
  *
  * @package PhpMyAdmin-test
  */
